fix: config 파일 중복 제거
