--- src/main/java/com/muud/config/WebConfig.java
@@ -1,19 +1,0 @@
-package com.muud.config;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.servlet.config.annotation.CorsRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-@Configuration
-public class WebConfig implements WebMvcConfigurer {
-
-    @Override
-    public void addCorsMappings(CorsRegistry registry) {
-        registry
-                .addMapping("/**")
-                .allowedHeaders("*")
-                .allowedOrigins("http://localhost:5173")
-                .allowedMethods("*")
-                .allowCredentials(true);
-    }
-}
