+++ src/main/java/com/muud/global/config/WebConfig.java
@@ -3,12 +3,22 @@ package com.muud.global.config;
 import com.muud.auth.jwt.JwtInterceptor;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.web.servlet.config.annotation.CorsRegistry;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 @RequiredArgsConstructor
 @Configuration
 public class WebConfig implements WebMvcConfigurer {
     private final JwtInterceptor jwtInterceptor;
+    @Override
+    public void addCorsMappings(CorsRegistry registry) {
+        registry
+                .addMapping("/**")
+                .allowedHeaders("*")
+                .allowedOrigins("http://localhost:5173")
+                .allowedMethods("*")
+                .allowCredentials(true);
+    }
 
     private String[] excludePathPatterns = {
             "/", "/auth", "/auth/**"
